refactor: move diff to pipeline file
