+++ pipeline.go
@@ -1,11 +1,32 @@
 package main
 
 import (
-	"fmt"
+	"strings"
+
+	log "github.com/sirupsen/logrus"
 )
+
+func diff(command string) []string {
+	log.Infof("Running diff command: %s", command)
+
+	split := strings.Split(command, " ")
+	cmd, args := split[0], split[1:]
+
+	output := executeCommand(cmd, args)
+
+	log.Debug("Output from diff:\n" + output)
+
+	return strings.Split(strings.TrimSpace(output), "\n")
+}
+
+func pipelinesToTrigger(diffCmd string) []string {
+	changedFiles := diff(diffCmd)
+
+	return changedFiles
+}
 
 func uploadPipeline() string {
 	msg := "uploading pipelines"
-	fmt.Println(msg)
+	log.Println(msg)
 	return msg
 }
